+++ src/main/java/com/anyang/maruni/domain/alertrule/application/analyzer/strategy/EmotionPatternAnalyzer.java
@@ -1,4 +1,4 @@
-package com.anyang.maruni.domain.alertrule.application.analyzer;
+package com.anyang.maruni.domain.alertrule.application.analyzer.strategy;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -7,6 +7,9 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Component;
 
+import com.anyang.maruni.domain.alertrule.application.analyzer.vo.AlertResult;
+import com.anyang.maruni.domain.alertrule.application.analyzer.vo.AnalysisContext;
+import com.anyang.maruni.domain.alertrule.application.analyzer.util.AnalyzerUtils;
 import com.anyang.maruni.domain.alertrule.application.config.AlertConfigurationProperties;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertLevel;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertType;
