+++ src/main/java/com/anyang/maruni/domain/alertrule/application/analyzer/vo/AlertResult.java
@@ -1,4 +1,4 @@
-package com.anyang.maruni.domain.alertrule.application.analyzer;
+package com.anyang.maruni.domain.alertrule.application.analyzer.vo;
 
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertLevel;
