+++ src/main/java/com/anyang/maruni/domain/alertrule/application/dto/request/AlertConditionDto.java
@@ -1,4 +1,4 @@
-package com.anyang.maruni.domain.alertrule.application.dto;
+package com.anyang.maruni.domain.alertrule.application.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Size;
