+++ src/main/java/com/anyang/maruni/domain/alertrule/application/dto/response/AlertHistoryResponseDto.java
@@ -1,4 +1,4 @@
-package com.anyang.maruni.domain.alertrule.application.dto;
+package com.anyang.maruni.domain.alertrule.application.dto.response;
 
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertHistory;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertLevel;
