+++ src/main/java/com/anyang/maruni/domain/alertrule/presentation/controller/AlertRuleController.java
@@ -14,13 +14,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.anyang.maruni.domain.alertrule.application.dto.AlertConditionDto;
-import com.anyang.maruni.domain.alertrule.application.dto.AlertDetectionResultDto;
-import com.anyang.maruni.domain.alertrule.application.dto.AlertHistoryResponseDto;
-import com.anyang.maruni.domain.alertrule.application.dto.AlertRuleCreateRequestDto;
-import com.anyang.maruni.domain.alertrule.application.dto.AlertRuleResponseDto;
-import com.anyang.maruni.domain.alertrule.application.dto.AlertRuleUpdateRequestDto;
-import com.anyang.maruni.domain.alertrule.application.service.AlertRuleService;
+import com.anyang.maruni.domain.alertrule.application.dto.request.AlertConditionDto;
+import com.anyang.maruni.domain.alertrule.application.dto.response.AlertDetectionResultDto;
+import com.anyang.maruni.domain.alertrule.application.dto.response.AlertHistoryResponseDto;
+import com.anyang.maruni.domain.alertrule.application.dto.request.AlertRuleCreateRequestDto;
+import com.anyang.maruni.domain.alertrule.application.dto.response.AlertRuleResponseDto;
+import com.anyang.maruni.domain.alertrule.application.dto.request.AlertRuleUpdateRequestDto;
+import com.anyang.maruni.domain.alertrule.application.service.core.AlertRuleService;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertRule;
 import com.anyang.maruni.domain.member.domain.entity.MemberEntity;
 import com.anyang.maruni.global.response.annotation.AutoApiResponse;
